requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.0
Django==3.0
django-crispy-forms==1.8.1
django-filter==2.2.0
djangorestframework==3.10.3
firebase-admin==3.2.1
gcloud==0.17.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.0
hiredis==1.0.1
httplib2==0.17.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jws==0.1.3
Markdown==3.1.1
msgpack==0.6.2
oauth2client==3.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.4.3
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2019.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
uritemplate==3.0.1
urllib3==1.25.7
zope.interface==4.7.1
